피드백 관련 API 권한 통합
